requirements.txt
joblib~=1.4.2
numpy~=2.1.1
pandas~=2.2.3
streamlit~=1.38.0
scikit-learn~=1.5.2
pillow~=10.4.0
matplotlib~=3.9.2
seaborn~=0.13.2